+++ project/source/game/net/Net.cpp
@@ -3302,14 +3302,14 @@ bool Game::ProcessControlMessageServer(BitStream& stream, PlayerInfo& info)
 					Error("Update server: Broken IDLE from %s.", info.name.c_str());
 					StreamError();
 				}
-				else if(index >= unit.data->idles->size())
+				else if(index >= unit.data->idles->anims.size())
 				{
 					Error("Update server: IDLE from %s, invalid animation index %u.", info.name.c_str(), index);
 					StreamError();
 				}
 				else
 				{
-					unit.mesh_inst->Play(unit.data->idles->at(index).c_str(), PLAY_ONCE, 0);
+					unit.mesh_inst->Play(unit.data->idles->anims[index].c_str(), PLAY_ONCE, 0);
 					unit.mesh_inst->groups[0].speed = 1.f;
 					unit.mesh_inst->frame_end_info = false;
 					unit.animation = ANI_IDLE;
@@ -6481,14 +6481,14 @@ bool Game::ProcessControlMessageClient(BitStream& stream, bool& exit_from_server
 						Error("Update client: IDLE, missing unit %d.", netid);
 						StreamError();
 					}
-					else if(animation_index >= unit->data->idles->size())
+					else if(animation_index >= unit->data->idles->anims.size())
 					{
-						Error("Update client: IDLE, invalid animation index %u (count %u).", animation_index, unit->data->idles->size());
+						Error("Update client: IDLE, invalid animation index %u (count %u).", animation_index, unit->data->idles->anims.size());
 						StreamError();
 					}
 					else
 					{
-						unit->mesh_inst->Play(unit->data->idles->at(animation_index).c_str(), PLAY_ONCE, 0);
+						unit->mesh_inst->Play(unit->data->idles->anims[animation_index].c_str(), PLAY_ONCE, 0);
 						unit->mesh_inst->groups[0].speed = 1.f;
 						unit->mesh_inst->frame_end_info = false;
 						unit->animation = ANI_IDLE;
