+++ project/source/game/stats/StatProfile.h
@@ -18,4 +18,7 @@ struct StatProfile
 	void Set(int level, UnitStats& stats) const { Set(level, stats.attrib, stats.skill); }
 	void SetForNew(int level, int* attribs, int* skills) const;
 	void SetForNew(int level, UnitStats& stats) const { Set(level, stats.attrib, stats.skill); }
+
+	static vector<StatProfile*> profiles;
+	static StatProfile* TryGet(const AnyString& id);
 };
